put back device-info-msg.h, without dependency on eProsima headers
