+++ include/librealsense2/dds/topics/device-info/device-info-topic-files.cmake
@@ -5,6 +5,7 @@ set(DEVICE_INFO_TOPIC_FILES
    "${CMAKE_CURRENT_LIST_DIR}/deviceInfo.h"
    "${CMAKE_CURRENT_LIST_DIR}/deviceInfo.cpp"
    "${CMAKE_CURRENT_LIST_DIR}/device-info-msg.h"
+   "${CMAKE_CURRENT_LIST_DIR}/device-info-msg.cpp"
    "${CMAKE_CURRENT_LIST_DIR}/deviceInfoTypeObject.h"
    "${CMAKE_CURRENT_LIST_DIR}/deviceInfoTypeObject.cpp"
    "${CMAKE_CURRENT_LIST_DIR}/deviceInfoPubSubTypes.h"
